+++ migrations.php
@@ -3,9 +3,14 @@
 declare(strict_types=1);
 
 return [
-    'name' => 'ShlinkMigrations',
-    'migrations_namespace' => 'ShlinkMigrations',
-    'table_name' => 'migrations',
-    'migrations_directory' => 'data/migrations',
+
+//    'name' => 'ShlinkMigrations',
+    'migrations_paths' => [
+        'ShlinkMigrations' => 'data/migrations',
+    ],
+    'table_storage' => [
+        'table_name' => 'migrations',
+    ],
     'custom_template' => 'data/migrations_template.txt',
+
 ];