+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 #### Changed
 
 * [#508](https://github.com/shlinkio/shlink/issues/508) Added mutation checks to database tests.
+* [#790](https://github.com/shlinkio/shlink/issues/790) Updated to doctrine/migrations v3.
 
 #### Deprecated